[ADD] Clase 9 - Java - 9.9 Función ejecutarOperacion y pruebas part 2

+++ Tecnicatura3Java/Leccion09/CalculadoraUTN/src/CalculadoraUTN.java
@@ -37,11 +37,11 @@ public class CalculadoraUTN {
     }// Fin de la funcion mostrarMenu
     private static void ejecutarOperacion(int operacion, Scanner entrada){
         System.out.print("Digite el valor para operando 1: ");
-        var operando1 = Integer.parseInt(entrada.nextLine());
+        var operando1 = Double.parseDouble(entrada.nextLine());
         System.out.print("Digite el valor para operando 2: ");
-        var operando2 = Integer.parseInt(entrada.nextLine());
+        var operando2 = Double.parseDouble(entrada.nextLine());
 
-        int resultado;
+        double resultado;
         switch (operacion){
             case 1 -> { // Suma
                 resultado = operando1 + operando2;
